Fix imports to avoid client-side errors
Fixes #370

Organize `server/utils.ts` functions to avoid client-side imports.

* Add `template/app/src/server/envUtils.ts` to contain the `requireNodeEnvVar` function.
* Remove `requireNodeEnvVar` function from `template/app/src/server/utils.ts`.
* Update imports of `requireNodeEnvVar` in `template/app/src/payment/plans.ts`, `template/app/src/payment/lemonSqueezy/paymentProcessor.ts`, `template/app/src/payment/stripe/paymentProcessor.ts`, `template/app/src/payment/stripe/stripeClient.ts`, `template/app/src/payment/lemonSqueezy/webhook.ts`, and `template/app/src/payment/stripe/webhook.ts` to import from `template/app/src/server/envUtils.ts`.
