+++ template/app/src/server/envUtils.ts
@@ -1,0 +1,8 @@
+export function requireNodeEnvVar(name: string): string {
+  const value = process.env[name];
+  if (value === undefined) {
+    throw new Error(`Env var ${name} is undefined`);
+  } else {
+    return value;
+  }
+}
