--- template/app/src/server/utils.ts
@@ -1,8 +1,0 @@
-export function requireNodeEnvVar(name: string): string {
-  const value = process.env[name];
-  if (value === undefined) {
-    throw new Error(`Env var ${name} is undefined`);
-  } else {
-    return value;
-  }
-}
